[Мета] Вынесены перечисления состояний из классов команд.
