+++ pet-impl/src/main/java/com/starbun/petproject1/command/CommandNames.java
@@ -2,5 +2,5 @@ package com.starbun.petproject1.command;
 
 public interface CommandNames {
   String COMMAND_START = "commandStart";
-  String COMMAND_DEBT = "commandStart";
+  String COMMAND_DEBT = "commandDebt";
 }
