+++ pet-impl/src/main/java/com/starbun/petproject1/command/debt/DebtKeyboardService.java
@@ -1,7 +1,6 @@
 package com.starbun.petproject1.command.debt;
 
 import com.starbun.petproject1.command.CommandStates;
-import com.starbun.petproject1.command.debt.DebtCommand.DebtCommandStates;
 import com.starbun.petproject1.dto.ButtonAction;
 import com.starbun.petproject1.exception.NoImplementationException;
 import com.starbun.petproject1.service.InlineKeyboardService;
